refactor(mjh): PR-gate compliance — component extraction, ExportFormat const, composer E2E
- ConversationHistory split: ConversationBubble / SectionDivider /
  PendingEchoBubbles own files; pure render helpers in
  conversation-turn-render.ts.
- InlineText extracted from markdown-preview.tsx.
- ExportFormat const module replaces "pdf"/"docx" magic strings in
  CompletePanel.
- E2E: route-mocked active session exercising the composer contract —
  always visible, clears on Enter, optimistic echo + thinking bubble,
  history renders the user's own words (no "Try something with:").
