+++ apps/myjobhunter/frontend/src/features/resume_refinement/export-format.ts
@@ -1,0 +1,11 @@
+/**
+ * Download formats for the completed resume. Values match the backend
+ * export endpoint's ``format`` query param
+ * (GET /resume-refinement/sessions/{id}/export?format=…).
+ */
+export const ExportFormat = {
+  PDF: "pdf",
+  DOCX: "docx",
+} as const;
+
+export type ExportFormat = (typeof ExportFormat)[keyof typeof ExportFormat];
